refactor: use entity instead of entity name in skin
