+++ packages/core/src/mesh/Skin.ts
@@ -1,24 +1,22 @@
 import { Matrix } from "@oasis-engine/math";
 import { EngineObject } from "../base/EngineObject";
-
-let skinID = 0;
+import { Entity } from "../Entity";
 
 /**
  * Mesh skin data, equal glTF skins define
  */
 export class Skin extends EngineObject {
   public inverseBindMatrices: Matrix[];
-  public joints: string[];
-  public skeleton: string;
+  public joints: Entity[];
+  public skeleton: Entity;
   /**
    * Contructor of skin
    * @param name - name
    */
-  constructor(name: string) {
+  constructor(public name: string) {
     super(null);
-
     this.inverseBindMatrices = []; // inverse bind matrix array, element type: gl-matrix.mat4
-    this.joints = []; // joints name array, element type: string
-    this.skeleton = "none"; // root bone name
+    this.joints = []; // joints entity array
+    this.skeleton = null; // root entity
   }
 }
